+++ lib/project_types/script/layers/infrastructure/api_clients/partners_proxy_api_client.rb
@@ -1,44 +1,7 @@
 module Script
   module Layers
     module Infrastructure
-      class ApiClients
-        def self.default_client(ctx, api_key)
-          if ENV["BYPASS_PARTNERS_PROXY"]
-            ScriptServiceApiClient.new(ctx, api_key)
-          else
-            PartnersProxyApiClient.new(ctx, api_key)
-          end
-        end
-
-        class ScriptServiceApiClient
-          LOCAL_INSTANCE_URL = "https://script-service.myshopify.io"
-
-          def initialize(ctx, api_key)
-            instance_url = script_service_url
-            @api = ShopifyCLI::API.new(
-              ctx: ctx,
-              url: "#{instance_url}/graphql",
-              token: { "APP_KEY" => api_key }.compact.to_json,
-              auth_header: "X-Shopify-Authenticated-Tokens"
-            )
-          end
-
-          def query(query_name, variables: {})
-            @api.query(query_name, variables: variables)
-          end
-
-          private
-
-          def script_service_url
-            if ::ShopifyCLI::Environment.use_spin_partners_instance?
-              "https://script-service.#{::ShopifyCLI::Environment.spin_url}"
-            else
-              LOCAL_INSTANCE_URL
-            end
-          end
-        end
-        private_constant(:ScriptServiceApiClient)
-
+      module ApiClients
         class PartnersProxyApiClient
           def initialize(ctx, api_key)
             @ctx = ctx
@@ -82,7 +45,6 @@ module Script
           end
           private_constant(:ShimAPI)
         end
-        private_constant(:PartnersProxyApiClient)
       end
     end
   end
